tratamentos de erros 404

+++ app/Http/Controllers/ClientsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Client;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 
 class ClientsController extends Controller
@@ -14,7 +15,11 @@ class ClientsController extends Controller
 
     public function show($id)
     {
-        return Client::find($id);
+        if (!($client = Client::find($id))) {
+            throw new ModelNotFoundException("Client requisitado não existe");
+        } else {
+            return $client;
+        }
     }
 
     public function store(Request $request)
@@ -26,21 +31,24 @@ class ClientsController extends Controller
 
     public function update(Request $request, $id)
     {
-        $client = Client::find($id);
+        if (!($client = Client::find($id))) {
+            throw new ModelNotFoundException("Client requisitado não existe");
+        } else {
+            $client->fill($request->all());
 
-        $client->fill($request->all());
+            $client->save();
 
-        $client->save();
-
-        return response()->json($client, 200);
+            return response()->json($client, 200);
+        }
     }
 
     public function destroy($id)
     {
-        $client = Client::find($id);
-
-        $client->delete();
-
-        return response()->json(true, 204);
+        if (!($client = Client::find($id))) {
+            throw new ModelNotFoundException("Client requisitado não existe");
+        } else {
+            $client->delete();
+            return response()->json("", 204);
+        }
     }
 }
